Fixed error with output in neuron
Problem was that the output had a weight of 0 thus no change in the
neuron ever got far.

Added console write lines in order to make perceptron trainer easier to
debug.

+++ AriesNeuroNet/Neurons/NeuronBase.cs
@@ -14,8 +14,8 @@ namespace AriesNeuroNet.Neurons
 
         // The input is only a list of ports
         public List<NeuronPort> inputs;
-        public int maxWeight { get; set; }
-        public int minWeight { get; set; }
+        public double maxWeight { get; set; }
+        public double minWeight { get; set; }
 
         // The output is also a port
         public NeuronPort output;
